Reset DSF code and add cofactors.
The DSF physics module has been reset to commit 4cb02891ad7f2d41a4fdf234a149cec0f2384016 when it was confirmed to be working. It will now be recoded from this point to include all the features that were present in the 0.6.0 release but maintaining correctness.

This commit includes the reading of cofactors from the configuration file and applying to the spin components.