Removed upSpace code.
Removed all references to the upSpace arrays which were to be used for spacial upsampling. This includes all the old commented code which was written for Gaussian convolution which is now done as a post processes on the data with a python script.

q-space dimensions are now stored in qDim not rDim.m