Refactor: JAMS argument is now full config filename
Previously the config file was specified by just the name (without .cfg) of the file which was used as a seed name for the output files. This is somewhat perverse and doesn't allow tab competition.

Now the full config filename is given and the path and extension are stripped in code to give the seed name.