Bugfix: Fix basis vectors to be columns of the basis matrix
Before, each basis vector was defined as a row of the unit cell matrix. This means that instead of doing the matrix multiplication with lattice vectors, the transpose multiplication had to be done. This is clearly a bad idea.

Now the basis vectors are read in as columns of the matrix. matmul can then be used with the unit cell matrix and lattice vectors.