Refactor: Massive refactoring
I have refactored the entire codebase into a more modular object oriented design.

Essentially, the solver class now runs the show. Attached to it are the physics and monitor classes which are called from the solver. Inheritance is now made more use of, where often a base class does the leg work of initialising from the config file and the specific implementation is in the class which derives the base.

There is still a lot of cleaning up and testing to be done.

I have also removed many of the more complex solvers which will be reimplemented in the future with the new system.