Remove spin initialization from CUDACONSTRAINEDMC
Previously I was initializing all spins to the constraint direction, but this is not always what we want to do (for example in an antiferromagnet). This can put us a long way from the initial energy minimum. Instead the user should initialize in the config file in the usual way, but probably using some wisdom to account for the constraint.