Hack to support older enterprise linux systems
Most enterprise systems have a fairly old version of GLIBC. This means if we compile on a newer system the binary is not compatible. The main differences are in the math '_finite' functions and also some terrible saga (http://www.win.tue.nl/~aeb/linux/misc/gcc-semibug.html) to do with memcpy. To avoid any libmath _finite functions we can use -fno-finite-math-only, which is also IEEE safe. I also added -fno-stack-protector as some googling suggests this can break compatibility as it adds an extra layer of hardening. Memcpy is more tricky and involves defining a wrapper to use the older memcpy version (which is actually the same as GLIBC_2.14). This follows advice from http://stackoverflow.com/questions/8823267/linking-against-older-symbol-version-in-a-so-file. So the binary is now compatible with older systems.

As a note, to check which symbols are referenced in the binary we can use 'objdump -x jams++ | grep GLIBC_2.14'