Calculate Skyrmion center of mass using PBC
The periodic boundaries make calculating the COM difficult in the case that the skyrmion wraps across a boundary. There is an algorithm (L. Bai and D. Breen, Journal of Graphics, GPU, and Game Tools 13, 53 (2008)) which deals with this by wrapping the easy of the 2D edges into a cylinder and calculating the COM from that.