Change MetropolisMC loop to select spins in order
Choosing random spins to update in the MetropolisMC loop causes a large performance hit (tested with clang++-7.0 on OSX).

Profiling with Instruments on OSX seems to suggest this is because the exchange calculation suffers a lot of cache misses due to the inability to prefect data.

Updating spins in sequential order is a valid method in the Metropolis method.

The code change simply replaces the random spin selection with choosing spins in order.

Doing this has provided ~4x speedup (using GdIG input).  The move acceptence rate has not been affected and the results are consistent with the random selection.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>