Create useful exchange neighbour list structure
Previously the interactions were directly inserted into the 3Nx3N sparsematrix. This meant that if we want to use the exchange neighbour list in other parts of the code it is not in an easily usable form.

I have refactored to create a structure which stores the neighbour list in a nice iteratable form where the outer dimension is of length num_spins and the inner dimension contains all of the interactions for each spin.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>