Created lattice members to calculate distances
The calculation of distances between different sites in the lattice can be complicated because of the potential for periodic boundary conditions. When there are periodic boundaries the minimum image convention must be applied and this occurs in the lattice space (not Cartesian space).

I've created three new members:
- minimum_image_fractional
- distance
- distance_sq

minimum_image_fractional: applies the minimum image convention to a vector which is already in fractional units of the lattice vectors (i.e. avoids casting into Cartesian space)

distance: returns the Cartesian distance between two points (i, j) on in the lattice, respecting the boundary conditions

distance_sq: returns the squared Cartesian distance between two points (i, j) on the lattice, respecting the boundary conditions

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>