Create lattice member to calculate neighbour lists
Sometimes we want a neighbour list based purely on distances between the lattice sites, not the exchange interactions. I've implemented this in the lattice class using a brute force method. It finds all of the neighbours for each site within a radius r_cutoff. Optionally if prune = true it will find only the nearest neighbours from that set.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>