Add NearTree constructor from a vector
The new constructor allows the NearTree to be constructed using a vector of nodes to be inserted into the tree.

There is also an optional argument 'randomize' which defaults to true. This then randomizes the order of the insertions. The lookup of neighbours is much faster with random insertions rather than inserting in a regular way over space.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>