Refactor of a lot of lattice code
The lattice code is quite messy, so I have begun to refactor it to make it easier to read and more usable. As part of this I have begun adding small structs to 'core/types.h' which packagae data nicely.

The purpose of the Lattice class is not to be very fast, but to provide information for other classes which can then create more optimal data structures to contain the information they need.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>