Change InteractionList in exchange to use maps
To calculate the 'bond' energy in a Wolff cluster algorithm I need to look up the Jij between spins without summing over all neighbours. This is not easy in a sparse matrix but can be done efficiently using the InteractionList if the inner dimension is a map where the key is the spin number and the value is the exchange tensor.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>