Correct exchange double counting and factors of half
When calculating the exchange interaction I was double counting because
of inserting both the site 'i' and it's neighbour 'j', but looping over
all i and j. This caused 2x interactions to be inserted. The sparse
matrix then sums the contributions.

Secondly, the energy calculation should include a factor of 0.5
from the Hamiltonian (H = - 1/2 J \sum S_i S_j).

Lastly, there was an erroneous factor of 2 in the percollation
probability for the Wolff cluster.

These results have been confirmed on 2016-01-06 for Metropolis
and Wolff Monte Carlo. For nearest neighbour simple cubic system
Tc = 1.44 J/kB, where J is the exchange specified in the exchange file
(i.e. no hidden factors of 2 or 0.5).