Correct structure factor delta_freq
The current calculation of the delta_freq is incorrect because the definition of max frequency is confused with the sample frequnecy.

The delta_freq is 1 / T = f_s / N, where T is the total aquisition time, f_s is the sample frequency (not the max frequency) and N is the number of samples.

I've also made the names of the frequency variables more consistent.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>