Unit cell positions can be read from config
The lattice class now checks if the "lattice.positions" config setting is a string or a list. If it is a string then it reads the file as before. If it is a list then it reads the atoms from the list in the config file.

The reason an external files was used before is because if the unitcell is huge (like an amorphous system) then libconfig would use huge amounts of memory.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>