Fix periodic lookup in metric for nonorthorhombic
The pervious code assigned the wrong spins as neighbours across
the periodic boundary if the unit cell was not orthorhmbic.

I've implemented the method from [W. Smith, CCP5 Information
Quarterly for Computer Simulation of Condensed Phases (1989)]
which is similar to what I was trying previously but turns the
whole super cell into an effective unit cell and defines the
positions then in fractional coordinates of this super unit cell.
This then makes the lookup trivial. Although there seems to be a
mod 1 missing in the paper, because if r_a == r_b the formula gives
a fractional distance of 1 but it should be zero.