Change geweke stat to discard first 10% of data
The first 10% of data can often be very far from equilibrium especially for trival MvT curves and low temperatures. By throwing away the first 10% of data and using the second 10% we should get much faster convergence in general. It also means the z-score should have more relavence because it is not scewed so much by very different initial states.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>