Add warmup for bose thermostat
I was using random numbers to seed initial values for the coloured noise process. This is incorrect as the zeta arrays should be seeded from distribution functions which in themselves are a pain to calculation.

Instead I seed the normally distributed eta values and initialize zeta to zero. Then doing a series of updates for the initials state to decay, the thermostat should be in a good state for simulations.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>