Change Geweke to use 0.1-0.2 for initial stats
In typical schenarios I start with m = 1 and this state decays to finite temperature. In this case, at higher temperatures it takes the initial state longer to decay and also the large value of m_0 compared to m_e creates a large bias for the initial state. I think it will be more efficient if the Geweke statistic is calculated using the second 10% which will gradually ignore the initial state.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>