Make lattice vector definitions consistent
The lattice vector matrix is a bit complicated, because each vector in principle could be defined on it’s own, but the unit cell matrix is the transpose of these vectors.

The previous behaviour required the vectors to be defined as the columns of the input unitcell. But this is not so intuitive because these don’t look like vectors in the config file. So now I have defined that the unitcell be read in a vectors and transposed after this.

One last complication is that spglib requires the input to be a set of vectors, the same as the jams configuration file now uses. So I must transpose the internal representation before giving it to spglib.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>