Implement metadynamics as a phyics/hamiltonian
I decided it’s best to implement the metadynamics as a physics/hamiltonian module pair. The Physics module is responsible for creating whatever metadynamics hamiltonian is requried and it then registers it with the solver through the normal hamiltonian register interface. I.e. they share a pointer to the same hamiltonian.

This is actually quite cool because the physics module can use the full MetadynamicsHamiltonian interface but the solver uses only the Hamiltonian interface.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>