Use the same basis definition as spglib
The different conventions for defining the basis vectors was causing a
lot of issues. I now use the same convention as spglib, i.e. the basis
vectors are columns of the matrix. For clarity, in the output I now give
both the a,b,c vectors and the matrix which is used.