Refactor sparse matrix to improve memory handling
Previously the memory usage when converting MAP to CSR in the sparse matrix was maximised because both the MAP and CSR data structures are allocated, with MAP only deallocated after the conversion is finished.

This refactor avoids allocating the CSR data structures at the start, but increases the size dynamically as the conversion proceeds. Also the size of the MAP is allowed to reduce by poping converted elements. It is then up to the system to resize the heap allocation for the vectors. This appears to work ok. If need be then the reallocation could be forced with the C++11 srink_to_fit.

There is some performance overhead to this dynamical resizing, but this is not too large and is greatly outweighted by the improved memory handling.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>