Googletest working with CUDA
Using googletest with CUDA has some issues. The main one is that you cannot create OBJECT libraries with cuda_add_library. The CUDA name mangling of object files means we cannot easily reuse the object files from the jams target with the jams_test target.

The solution for now is to rebuild all of the object files for the jams_test target.

I also had to spinoff the jams main() function into a new source file so that jams_test can have it’s own main.

This is quite hackish at the moment and maybe needs revisiting in the future.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>