Use switch statement in Vec operator[]
This fixes issue #6:

Using the square brackets operator [] sometimes gives incorrect results.

The results are usually correct but occasionally there is some non-specific inconsistency. Inserting other code in proximity, such as print statements can effect the result.

The bug originates from an assumption that POD data members are stored in consecutive memory addresses. The x,y,z data members were then accessed as a memory offset.

It seems that before C++11 this was the case. Now the allocation rules have been further clarified in the standard. Data members will have consecutively higher memory addresses but the compiler is free to pad the addresses. Therefore we can't use the offset trick anymore.

The simplest fix is to use a switch statement in the [] operator. This might reduce performance and in the future I should consider whether to do a more serious rewrite.

I also had to remove code which used the begin() and end() methods which now no longer can be used.

Signed-off-by: Joseph Barker <joseph.barker@imr.tohoku.ac.jp>