additional unit tests for fft dipole strategy
The fft method requires extensive testing because the calculation
must be performed for each pair of positions in the unit cell.
The same tests as the bruteforce method must be passed but we also
add some cross checks where the bruteforce and fft methods are
compared for the case of multiple atoms in the unit cell.

The 'RAND' cases use randomized spin configurations. These are
useful because they randomize the Fourier components. Most of the
analytical tests are pathological because only the k=0 Fourier
components exist.