refactor dipole fft for multiple atoms in unit cell
When there are multiple atoms in the unit cell we must calculate
all of the interaction tensors for the different pairings. There
are some symmetries, but we ignore that for now.

No phase factors are required because they are automatically
included by using r_j - r_i in real space.