fix #12: dipole FFT calculation in hexagonal systems
In the minimum image distance class we directly use the unitcell *matrix* but this is the transpose of the actual unit cell vectors. Therefore when we multiply by the size of the lattice to make the supercell matrix we should use the ‘j’ index of the size array.

Fixing this then solves the issues with the dipole FFT tensor.