fix calculation of inverse unitcell
After fixing the matrix inverse in commit e42c365 (which was actually the transpose of the inverse) the inverse unit cell calculation was now broken. This is because we need the transpose after the inverse (see https://en.wikipedia.org/wiki/Reciprocal_lattice or calculate using cross products for reciprocal vectors).

This commit fixes the inverse unit cell and puts it into a function to be consistent across the code base.