revert to complex to complex transform
Previously I used a real to complex transform with the spin array in place. However I actually need to apply to spin transformations, so this already means out of place is needed. On top of this, it is much simpler to look up brillouin zone indicies if I don’t have the last half dimension.