Basic impurity code using random seed in input
The code supports vacancies where the length of the spin is zero. This is done by checking for zero spins whenever we divide by the norm.

Using the same seed in the config file for the impurity generation will give the same atomic arrangement. This means we can perform different simulation stages with the same setup without having to store the lattice on disk.