Rename abs to norm for Vec3
The logical use of abs() in programming is to take the positive value. This can cause confusion with vectors where mathematically we often use abs to mean the length of the vector.

To avoid abiguity I’ve chosen to rename taking the length of the vector as norm() and use abs() now as making the components of the vector positive.