Move all FFTW code into interface header
This is important because of an incompatibility between come versions of CUDA and fftw/mkl which does not properly detect the NVCC compiler. It needs to do this to disable quad precision routines which rely on __float128 which NVCC does not provice.

I’ve used a hack during the inclusion of fftw3.h to avoid this issue, but all code must now include the interface header not fftw3.h directly.