Fix #32 Segfault using MKL FFTW and release build
The operator= for multiarray always causes a resize and therefore reallocation of the underlying memory. This causes the pointer which is used in the FFTW planning to change and become invalid.

Now we check for the case of the two containers being the same size and avoid reallocating if possible. I think this needs a more general solution where the underlying synched_memory only copies the parts that need copying (i.e. GPU or CPU memory).