Remove explicit use of H5Cpp library
The H5 built with C++ and MPI is not an officially supported configuration and therefore not available on many clusters.

The HighFive C++ header library gives a much better interface for writing H5 and also does not use the H5 C++ bindings.

I have replaced all explicit H5Cpp calls with HighFive. We still find HDF5 with cmake so that we can pickup the library info for version.h.