Add config option for random spin order
The Metropolis Monte Carlo used to go through the spins in order with each trial move. On some systems this gave a 4x speedup.

In Mai Kameda’s skyrmion work Kei Yamamoto said he saw some strange skew to the figures. There’s a small possibility this could be from striding through the lattice systematically.

Now the spins are selected at random by default and the option to be random or ordered can be toggled in the configuration file.