Rewrite of SparseMatrix
Sparse matricies are now built with a builder class which emits a chosen format of sparse matrix.

Sparse matrix - dense vector multiplication is now done byt the SparseMatrix class whether on the CPU or GPU. All of the allocation and copying to the GPU is handled internally byt the SparseMatrix class.