Refactor InteractionList with flat vectors
This is not the fastest or most elagent way to make an interaction list, but it allows a single class to be used for interaction lists of any number of indicies (e.g. 2, 3, 4…).

Memory use is minimised by using a VectorSet mapping so only unique interaction values are stored. This are looked up by an integer key.