Fix rotation matrices are incorrect if phi is not zero
The logic of the rotations was incorrect and because 3D rotations do not commute this caused the total rotation matrix to be incorrect.

When doing the inverse rotation (i.e. from where the constraint vector is along z to the true cartesian axes) we need to first rotate the system from z-constraint to z-cartesian (using an Ry matrix) and then rotate in the xy plane (using an Rz matrix) to get the correct phi angle.