Add default to align the spins to the constraint
I added a configurable setting which will automatically align the spins to the constraint during initialisation.

This means we don’t need to bother working out a theta and phi for the spins as long as the transform matrix is correct.

This code accounts for the different values of the moment, i.e. it aligns the transformed magnetisation (which is the correct thing to do).