Create singleton for global handles
The new singleton ‘Jams’ contains all of the CUDA handles, the random number generator and a check for the CPU or GPU mode. This allows us to check if we should us CudaMalloc or not in the memory allocation (for pinned memory).

Overall this means we can run even when no GPU is present.