Fix bug in matrix symmetry check
std::lower_bound check for the first element not greater than or equal to the argument. This means the value can be less than but not equal to the argument, but we are looking for a range equal to the argument.

The fix is just to check if the found lower bound is indeed equal.