Fix fractional coordinate shift at 1
Fractional coordinates must in the range 0 <= x < 1. The current conversion and shift allows coordinates of 1, but these should be mapped back to 0 because 1 is within the next cell.

We now check if any coordinate is very close to 1 and map it back to 0 if it is. we do the same in the validation function too.