Fix incorrect formula in rotation_matrix_between_vectors
Rodrigues formula was missing the normalisation on the outer_product leading to rotation matricies with a determinant not equal to one.

I’ve fixed this added some assert statements to ensure the rotation matrix is always valid when we are debugging.

I also changed from copying the vectors to using references.