Add DipoleNeighbourList Hamiltonian
This is a compromise between using a NearTree (low memory but slower because of neighbour list lookup) and a sparse tensor (fast to calculate but very large memory use).

In testing it seems this new Hamiltonian may be the fastest for Monte Carlo, presumably because the sparse matrix tensor has to do 3 row multiplications which don’t have optimal memory access.