Add method to optimise the NearTree in the lattice class
The NearTree can be made more optimal by removing mirrored atoms which cannot be reached (are not within the maximum interaction radius of any atom in the central supercell). But generating this is costly for large systems. So we make it an optional function which can be envoked if it’s really needed.