Fix rotation_matrix_between_vectors
I had previously missed a norm from the formula but clearly not tested after fixing with the division of the norm. This norm is zero if cross(a,b) == 0 (i.e. the vectors point in the same direction).

It seems there’s no way to avoid using an if statement to check for this case, so I have now gone back to the simplest statement for Rodrigues’s formula and catch cross(a,b) == 0 with an if statement.

Note that the unit tests already exist for this and did indeed fail before the fix.