Support multiple config files and patches as args
Any non-flag arguments are now treated as config strings. They are read in order. If the string corresponds to a valid file, it is read, otherwise it is treated as a string containing settings.

The settings are combined into a single config where later settings will be added or overwrite earlier settings.

The simulation name is now taken as the name of the first of the config files. If there is no config file (for example if all of the settings are given in a string) the simulation name defaults to “jams”.