Fix removal of elastic scattering contribution
There was a mistake in my derivation and this contribution was missing in the SpectrumBaseMonitor and incorrect in the NeutronScatteringNoLatticeMonitor.

The elastic contribution needs to be removed before the convolution in time. In my derivation I had missed including it in the convolution theorem and simply subtracted with a delta(omega) outside of the convolution which is incorrect.