Fix dipole symmetry check (#77)
* Add explanatory comments

* Fix dipole symmetry check

We previously checked the symmetry for generated interactions between each combination of positions within the unit cell. However, multiple positions can have the same material which might be one of the symmetric points. So the check can fail even because some of the symmetric points are generated with a different pos_i, pos_j combination.

We therefore have to generate all of the points for a given pos_i before checking that this set is symmetry complete.