Read spglib version at runtime (#81)
We previously assumed that the JAMS_SPGLIB_VERSION in CMakeLists is correct and used that for the version number. But there is a bug so the automatic downloading is not updating so the version printed in the jams header is still the old version, but we have set JAMS_SPGLIB_VERSION to a new version.

We fix this by using spglibs functions to get the version string.