Add build flag to toggle external project updates
All the projects which were being added using ExternalProject_add had the argument “UPDATE_DISCONNECTED 1” which means they didn’t check for updates when cmake runs. Hence changing the version number in the CMakeLists.txt would not cause a new version to be pulled and built.

Now we use the directory level EP_UPDATE_DISCONNECTED flag to toggle updates on or off using a build flag JAMS_BUILD_OFFLINE. By default we now choose to always check for updates when cmake runs. We are always specifying the version we want (with a git tag or hash) so this will only actually update if we change the version number.