Refactor minimum image functions
Move the minimum image code into stand alone functions which are not strongly coupled to the Cell class.

The Smith’s algorithm function has been improved so it performs the check for whether the result is guaranteed unique.

The bruteforce has also been refactored to ensure it always checks sufficient offset distances to find the shortest distance even for perversely skewed cells.

A general function is now available which will first try Smith’s algorithm and then fall back to bruteforce if needed.