Fix dipole bruteforce cutoff for difficult cells
Here we have implemented the new minimum image functions but the real fix is the use of definately_greater_than rather than ‘>’ for the floating point comparison of |r_ij| with the cutoff radius. Without the safe floating point comparison neighbours are missed when the unit cell has non-trivial vectors.