Fix bruteforce dipole hamiltonian for awkward unit cells (#87)
* Fix dipole tests use of global config

The global config was refactored to unique_ptr a while ago but the mocking code has not been updated.

* Fix test input functions for missing commas

Vec3 output used to include commas but does not anymore so we must include manually to avoid a syntax error in libconfig.

* Add dipole tests between methods

Compares methods for a simple cubic and non-cubic lattice to test if all methods give the same results.

* Include Cell test again

* Refactor minimum image functions

Move the minimum image code into stand alone functions which are not strongly coupled to the Cell class.

The Smith’s algorithm function has been improved so it performs the check for whether the result is guaranteed unique.

The bruteforce has also been refactored to ensure it always checks sufficient offset distances to find the shortest distance even for perversely skewed cells.

A general function is now available which will first try Smith’s algorithm and then fall back to bruteforce if needed.

* Fix dipole bruteforce cutoff for difficult cells

Here we have implemented the new minimum image functions but the real fix is the use of definately_greater_than rather than ‘>’ for the floating point comparison of |r_ij| with the cutoff radius. Without the safe floating point comparison neighbours are missed when the unit cell has non-trivial vectors.