Refactor metropolis for virtual members
To allow solvers to dervice from the metropolis solver we’re making some of the core elements virtual. Hence we can end up with “X is a type of Metropolis Monte Carlo solver”. Possibly this needs more extensive refactoring in the future such that Monte Carlo solvers are a general type.