Use unit vector length to determine if we zero
Previously we checked if each component of the unit vector was approximately zero to decide if we should return a zero vector instead of trying to normalise the vector.

Now we use the length instead which makes more intuitive sense. We also now use a much smaller epslion (DBL_EPSILON) because the zero vector should almost only ever be used when the vector has been explicitly set to zero.