Ensure spin vector is unit or zero when initialising the lattice (#94)
* Normalise spin vector upon reading config

* Normalise spin vector upon reading config

* Use unit vector length to determine if we zero

Previously we checked if each component of the unit vector was approximately zero to decide if we should return a zero vector instead of trying to normalise the vector.

Now we use the length instead which makes more intuitive sense. We also now use a much smaller epslion (DBL_EPSILON) because the zero vector should almost only ever be used when the vector has been explicitly set to zero.

* Clarified the logic of initial spin setup

I’ve reorded and cleaned up the code for initialising the spin. We normalise the length of the spin as the final operation to ensure whatever has happened the spin vectors in the global array are unit vectors or zero vectors.

Co-authored-by: Joseph Barker <j.barker@leeds.ac.uk>