fixed all errors from the single CV, added extra flexibility from config file
Added a bool “tempering” which is passed from the config file which switch on and off the tempering (def = false => no tempered) acting on “magnetisation_cd”)

Added a parameter for the “gaussian_deposition_stride_” passed from the confiq file to control the deposition rate and output on “cpu_metadynamics_metropolis_solver”)

simplified the output, now I output 2 files : 1. including the potential, 2. metadynamics_parameters (iteration, gaussian_amplitude, energy_barrier). Also the name of the file names indicate if the results are for “tempered” or “plain” metadynamics