Move neartree out of lattice class
Previously we always generated a neartree inside the lattice class for neighbour lookups by other classes. For large systems the creation time of the this neartree can be quite slow and is often not needed if no Hamiltonian uses them. It is also often the case that Hamiltonians only need to use a neartree during initialisation after which the memory could be released.

There was also a method to try and optimise the neartree but this was slow that it was barely usable.

I’ve now refactored it so the neartree is wrapped in a class “InteractionNearTree” which can be used easily in each Hamiltonian that requires a neartree. This class takes a cutoff radius and will only create a minimum image halo as far as this radius. This massively reduces the memory requirements for short ranged interactions.