Fix empty cache
This is a nasty hack which probably needs writing in a better way. Basically this class gets constructed earlier than we do some types of spin initialisation in JAMS meaning the cached centre of mass was incorrect until the cache was first updated.

The hack just makes sure we calculate the cached value the first time it is used. Probably some proper caching logic is neater.