Fix near tree neighbour list bug (#109)
* Fix: failure to propagate floating point epsilon

In this recursive function call we don't include the optional parameter epsilon so it uses the value of std::numeric_limits<double>::epsilon() which may be much smaller than we passed in. This causes issues such as missing neighbours in complex unit cells.

* Use definately_greater_than consistently

Change to always using the same comparitor for consistency. definately_greater_than is the one we use in bruteforce comparisons too.

* Remove default floating point epsilons

Remove all default values such that the user always must think of an appropriate value. This also helps to avoid propagating incorrect default values.

* Fix: Find in annulus to avoid self as a neighbour

* Compare squared distances with squared epsilons

* Add google tests to github action

* Fix SyncedMemory constructor overload resolution

Previously there where two constructors which could be overloaded in the same way for example if the type was int. Not we use enable_if to enable the constructor from iterators only if the types are truely iterators.