Change internal units
Within the code we now use the following units:

- time: picoseconds (ps)
- frequency: terahertz (THz)
- energy: millielectron volts (meV)
- field: Tesla (T)

This gives equations where physical constants are all reasonably close to 1 so
that the implementation in code can be written exactly the same as the maths on
paper (i.e. there are no arbitrary normalised units). We avoid Hartree atomic
units because the scales are quite a bit smaller than the relevant spin dynamic
scales and there's a high potential for bugs, for example because moments are
multiples of $2\mu_B$.