P29 lorentzian noise (#110)
* Adds Lorentzian quantum noise thermostat and solver (Anders, arXiv:2009.00600v2 (2021), https://arxiv.org/abs/2009.00600)

* Implements fourth order **stochastic** Runge-Kutta for Bose thermostat (Honeycutt, Phys. Rev. A 45, 604 (1992), https://dx.doi.org/10.1103/physreva.45.604)

* Fixed inconsistent use of definately_greater_than and std::numeric_limits<double>::epsilon()

* Adds google tests to github action

* Fixes SyncedMemory constructor overload resolution

* Disables dipole tensor tests, these are too slow for github automated testing. Refactor in future.

* Changes internal units

Within the code we now use the following units:

- time: picoseconds (ps)
- frequency: terahertz (THz)
- energy: millielectron volts (meV)
- field: Tesla (T)

This gives equations where physical constants are all reasonably close to 1 so
that the implementation in code can be written exactly the same as the maths on
paper (i.e. there are no arbitrary normalised units). We avoid Hartree atomic
units because the scales are quite a bit smaller than the relevant spin dynamic
scales and there's a high potential for bugs, for example because moments are
multiples of $2\mu_B$.

* Implements a CUDA fourth order Runge-Kutta solver for the LLG

* Removes Solver inherentance from Base

* Adds a macro to define solvers

* Removes destructor from Solver class

* Enable workflow_dispatch on github actions