Fix units of Gilbert damping in Lorentzian thermostat
In arXiv:2009.00600v2 Janet uses eta_G for the Gilbert damping, but this is a **dimensionful** Gilbert damping (implied by Eq. (1) in the paper and also explicitly mentioned). In JAMS alpha is the dimensionless Gilbert damping. The difference is eta_G = alpha / (mu_s * gamma). It's important that we convert here to get the scaling of the noice correct (i.e. it converts Janet's equations into the JAMS convention).