Fix incorrect loop limit
In the maths Janet truncates the sum and N_trunc which we would naturally do as N_trunc + 1 in C++. However the limit here is not consistent with the size of the arrays (filter and rands). The choice of the size of num_trunc is arbitrary so we simply change the limit on the sum.