Add setting to toggle random spin selection in Monte Carlo
We used to always generate random indicies to choose a spin to move, but it may be beneficial to go in order, for example in metadynamics to ensure every spin is definately tried every MC step. So we add this as a toggle setting select_spins_by_random in the solver.