Refactor metadynamics collective variables
Collective variables are now implemented as a very simple class which calculates a single value. The potential is now defined and manipulated in the MetadynamicsPotential class which holds the CollectiveVariable pointers.

Most CollectiveVariables are expensive so we have a derived caching class which helps to implement caching for Monte Carlo based solvers.