Generalising potential class to add DeathBC
two bools are introduced:
death_bc_passed -> true is the DeathBC is passed through the config file -> used to go into the if-statement to check the DeathBC.
death_bc_ -> if this is also true a signal will br sent to the solver to exit the simulation.

In order to check for the DeathBC Conditions:
I had to  store the range min&max for each collective variable in a vector, split the cvars_bc into upper upper and lower_bc . Introduce a bool function to check if the threshold for the DeathBC is its exceeded or is equal too.

Next step implement the changes, write a function which exit the simulation and test