Multidimensional collective variables (#116)
* Fix mistake in interpolation formula.

* Add setting to toggle random spin selection in Monte Carlo

We used to always generate random indicies to choose a spin to move, but it may be beneficial to go in order, for example in metadynamics to ensure every spin is definately tried every MC step. So we add this as a toggle setting select_spins_by_random in the solver.

* Refactor metadynamics collective variables

Collective variables are now implemented as a very simple class which calculates a single value. The potential is now defined and manipulated in the MetadynamicsPotential class which holds the CollectiveVariable pointers.

Most CollectiveVariables are expensive so we have a derived caching class which helps to implement caching for Monte Carlo based solvers.

* Output collective variables as we insert gaussians

* Add setting to delay the start of metadynamics

This allows us to run normal monte carlo for some iterations to equilibrate before starting the metadynamics process.