Amorphous spectrum (#118)
* Fixed neutron-scattering-no-lattice monitor

There were two main issues, one is that we cannot use the faster method for calcuting the spectrum where we multiply S_i with exp(-Q.R_i) because in the end R_ij can be different depending on whether a periodic boundary is crossed or not.

The second issue is that in the special fourier transform the zero mode was being counted twice which was producing the bands at constant frequency. Normally we need i,j and j,i but for i=j we should only count the mode once.