Fixed and Tested Death Boundary
2 extra parameters are now used:
lower_death_boundary
upper_death_boundary

Since, using the min-max range for topological charge as the boundaries creates a major issue. The value highewr or lower than the limits is interpolated in order to fit in the physical space; thus the stop_signal() is never trigered and the topological calculation seems wronf for these case.

The simulation in now stopped istanlty. --- no need for that since even for lower values than the boundary threshold can be in the physical space of the collective variable.