Append metadynamics potentials
We now open the metadynamics potential file in the solver class so we can append rather than overwrite.

We also calculate the tempering scaling factor and output the free energy so that this scaling doesn't need to be done as a post process (which requires knowing the temperature and temperaring bias temperature).