Remove random cell testing for minimum image
This often gave random failures during testing because random cells could be too skew for the explicit search depth we gave to the bruteforce method.

It's now rewritten as a parameterised test where we can give different cells.

I've also given a full list of positions many of which are symmetry points worth checking. Of particular note are degenerate points (with respect to the minimum image) such as 0.5,0.5,0.5 which showed that Smith's method and Bruteforce pick different vectors for the same r_ij.